+++ MahApps.Metro/Controls/DateTimePicker.cs
@@ -1,125 +1,48 @@
 ﻿namespace MahApps.Metro.Controls
 {
     using System;
-    using System.Collections.Generic;
-    using System.Collections.ObjectModel;
     using System.ComponentModel;
-    using System.Globalization;
-    using System.Linq;
     using System.Windows;
     using System.Windows.Controls;
-    using System.Windows.Controls.Primitives;
-    using System.Windows.Data;
     using System.Windows.Markup;
 
     /// <summary>
     ///     Represents a control that allows the user to select a date and a time.
     /// </summary>
-    [TemplatePart(Name = ElementButton, Type = typeof(Button))]
-    [TemplatePart(Name = ElementHourHand, Type = typeof(UIElement))]
-    [TemplatePart(Name = ElementHourPicker, Type = typeof(Selector))]
-    [TemplatePart(Name = ElementMinuteHand, Type = typeof(UIElement))]
-    [TemplatePart(Name = ElementSecondHand, Type = typeof(UIElement))]
-    [TemplatePart(Name = ElementSecondPicker, Type = typeof(Selector))]
-    [TemplatePart(Name = ElementMinutePicker, Type = typeof(Selector))]
-    [TemplatePart(Name = ElementAmPmSwitcher, Type = typeof(Selector))]
+    [TemplatePart(Name = ElementCalendar, Type = typeof(System.Windows.Controls.Calendar))]
     [DefaultEvent("SelectedDateChanged")]
-    public class DateTimePicker : Control
+    public class DateTimePicker : TimePartPickerBase<DateTime?>
     {
-        public static readonly DependencyProperty CultureProperty = DependencyProperty.Register(
-            "Culture",
-            typeof(CultureInfo),
-            typeof(DateTimePicker),
-            new PropertyMetadata(null, OnCultureChanged));
         public static readonly DependencyProperty DisplayDateEndProperty = DatePicker.DisplayDateEndProperty.AddOwner(typeof(DateTimePicker));
         public static readonly DependencyProperty DisplayDateProperty = DatePicker.DisplayDateProperty.AddOwner(typeof(DateTimePicker));
         public static readonly DependencyProperty DisplayDateStartProperty = DatePicker.DisplayDateStartProperty.AddOwner(typeof(DateTimePicker));
         public static readonly DependencyProperty FirstDayOfWeekProperty = DatePicker.FirstDayOfWeekProperty.AddOwner(typeof(DateTimePicker));
-        public static readonly DependencyProperty HandVisibilityProperty = DependencyProperty.Register(
-            "HandVisibility",
-            typeof(TimePartVisibility),
-            typeof(DateTimePicker),
-            new PropertyMetadata(TimePartVisibility.All, OnHandVisibilityChanged));
+
         public static readonly DependencyProperty IsClockVisibleProperty = DependencyProperty.Register(
             "IsClockVisible",
             typeof(bool),
             typeof(DateTimePicker),
             new PropertyMetadata(true, OnClockVisibilityChanged));
-        public static readonly DependencyProperty IsDropDownOpenProperty = DatePicker.IsDropDownOpenProperty.AddOwner(typeof(DateTimePicker), new PropertyMetadata(default(bool), OnIsDropDownOpenChanged));
-        public static readonly DependencyProperty IsReadOnlyProperty = DependencyProperty.Register(
-            "IsReadOnly",
-            typeof(bool),
-            typeof(DateTimePicker),
-            new PropertyMetadata(default(bool)));
+      
         public static readonly DependencyProperty IsTodayHighlightedProperty = DatePicker.IsTodayHighlightedProperty.AddOwner(typeof(DateTimePicker));
         public static readonly DependencyProperty OrientationProperty = DependencyProperty.Register(
             "Orientation",
             typeof(Orientation),
             typeof(DateTimePicker),
             new PropertyMetadata(Orientation.Horizontal, null, CoerceOrientation));
-        public static readonly DependencyProperty PickerVisibilityProperty = DependencyProperty.Register(
-            "PickerVisibility",
-            typeof(TimePartVisibility),
-            typeof(DateTimePicker),
-            new PropertyMetadata(TimePartVisibility.All, OnPickerVisibilityChanged));
         public static readonly RoutedEvent SelectedDateChangedEvent = DatePicker.SelectedDateChangedEvent.AddOwner(typeof(DateTimePicker));
         public static readonly DependencyProperty SelectedDateProperty = DatePicker.SelectedDateProperty.AddOwner(typeof(DateTimePicker), new PropertyMetadata(OnSelectedDateChanged));
-        public static readonly DependencyProperty SourceHoursProperty = DependencyProperty.Register(
-            "SourceHours",
-            typeof(ICollection<int>),
-            typeof(DateTimePicker),
-            new FrameworkPropertyMetadata(Enumerable.Range(0, 24).ToList(), FrameworkPropertyMetadataOptions.BindsTwoWayByDefault, null, CoerceSourceHours));
-        public static readonly DependencyProperty SourceMinutesProperty = DependencyProperty.Register(
-            "SourceMinutes",
-            typeof(ICollection<int>),
-            typeof(DateTimePicker),
-            new FrameworkPropertyMetadata(Enumerable.Range(0, 60).ToList(), FrameworkPropertyMetadataOptions.BindsTwoWayByDefault, null, CoerceSource60));
-        public static readonly DependencyProperty SourceSecondsProperty = DependencyProperty.Register(
-            "SourceSeconds",
-            typeof(ICollection<int>),
-            typeof(DateTimePicker),
-            new FrameworkPropertyMetadata(Enumerable.Range(0, 60).ToList(), FrameworkPropertyMetadataOptions.BindsTwoWayByDefault, null, CoerceSource60));
-        private const string ElementAmPmSwitcher = "PART_AmPmSwitcher";
-        private const string ElementButton = "PART_Button";
+      
         private const string ElementCalendar = "PART_Calendar";
-        private const string ElementHourHand = "PART_HourHand";
-        private const string ElementHourPicker = "PART_HourPicker";
-        private const string ElementMinuteHand = "PART_MinuteHand";
-        private const string ElementMinutePicker = "PART_MinutePicker";
-        private const string ElementPopup = "PART_Popup";
-        private const string ElementSecondHand = "PART_SecondHand";
-        private const string ElementSecondPicker = "PART_SecondPicker";
-        private Selector _ampmSwitcher;
-        private Button _button;
         private System.Windows.Controls.Calendar _calendar;
-        private bool _deactivateRangeBaseEvent;
-        private UIElement _hourHand;
-        private Selector _hourInput;
-        private UIElement _minuteHand;
-        private Selector _minuteInput;
-        private Popup _popup;
-        private UIElement _secondHand;
-        private Selector _secondInput;
-        private TimeSpan? _timeOfDay;
-        private bool _timeOfDayChanged;
 
         static DateTimePicker()
         {
             DefaultStyleKeyProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(typeof(DateTimePicker)));
-            VerticalContentAlignmentProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(VerticalAlignment.Center));
-            HorizontalContentAlignmentProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(HorizontalAlignment.Right));
-            LanguageProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(OnCultureChanged));
+            //VerticalContentAlignmentProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(VerticalAlignment.Center));
+            //HorizontalContentAlignmentProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(HorizontalAlignment.Right));
+            //LanguageProperty.OverrideMetadata(typeof(DateTimePicker), new FrameworkPropertyMetadata(OnCultureChanged));
         }
-
-        /// <summary>
-        ///     Occurs when the drop-down date-time-picker is closed.
-        /// </summary>
-        public event RoutedEventHandler Closed;
-
-        /// <summary>
-        ///     Occurs when the drop-down date-time-picker is opened.
-        /// </summary>
-        public event RoutedEventHandler Opened;
 
         /// <summary>
         ///     Occurs when the <see cref="SelectedDate" /> property is changed.
@@ -128,17 +51,6 @@
         {
             add { AddHandler(SelectedDateChangedEvent, value); }
             remove { RemoveHandler(SelectedDateChangedEvent, value); }
-        }
-
-        /// <summary>
-        ///     Gets or sets a value indicating the culture to be used in string formatting operations.
-        /// </summary>
-        [Category("Behavior")]
-        [DefaultValue(null)]
-        public CultureInfo Culture
-        {
-            get { return (CultureInfo)GetValue(CultureProperty); }
-            set { SetValue(CultureProperty, value); }
         }
 
         /// <summary>
@@ -187,20 +99,6 @@
         }
 
         /// <summary>
-        ///     Gets or sets a value indicating the visibility of the clock hands in the user interface (UI).
-        /// </summary>
-        /// <returns>
-        ///     The visibility definition of the clock hands. The default is <see cref="TimePartVisibility.All" />.
-        /// </returns>
-        [Category("Appearance")]
-        [DefaultValue(TimePartVisibility.All)]
-        public TimePartVisibility HandVisibility
-        {
-            get { return (TimePartVisibility)GetValue(HandVisibilityProperty); }
-            set { SetValue(HandVisibilityProperty, value); }
-        }
-
-        /// <summary>
         ///     Gets or sets a value indicating whether the clock of this control is visible in the user interface (UI). This is a
         ///     dependency property.
         /// </summary>
@@ -218,38 +116,7 @@
             set { SetValue(IsClockVisibleProperty, value); }
         }
 
-        /// <summary>
-        ///     Gets or sets a value that indicates whether the drop-down for a <see cref="DateTimePicker" /> box is currently
-        ///     open.
-        /// </summary>
-        /// <returns>true if the drop-down is open; otherwise, false. The default is false.</returns>
-        public bool IsDropDownOpen
-        {
-            get { return (bool)GetValue(IsDropDownOpenProperty); }
-            set { SetValue(IsDropDownOpenProperty, value); }
-        }
 
-        /// <summary>
-        /// Gets a value that indicates if the <see cref="DateTimeFormatInfo.AMDesignator"/> that is specified by the <see cref="CultureInfo"/> 
-        /// set by the <see cref="Culture"/> (<see cref="FrameworkElement.Language"/> if null) has not a value.
-        /// </summary>
-        public bool IsMilitaryTime
-        {
-            get { return string.IsNullOrEmpty(SpecificCultureInfo.DateTimeFormat.AMDesignator); }
-        }
-
-        /// <summary>
-        ///     Gets or sets a value that enables visible-only mode, in which the contents of the <see cref="DateTimePicker" /> are
-        ///     visible but not editable.
-        /// </summary>
-        /// <returns>
-        ///     true if the <see cref="DateTimePicker" /> is read-only; otherwise, false. The default is false.
-        /// </returns>
-        public bool IsReadOnly
-        {
-            get { return (bool)GetValue(IsReadOnlyProperty); }
-            set { SetValue(IsReadOnlyProperty, value); }
-        }
 
         /// <summary>
         ///     Gets or sets a value that indicates whether the current date will be highlighted.
@@ -276,20 +143,6 @@
         }
 
         /// <summary>
-        ///     Gets or sets a value indicating the visibility of the selectable date-time-parts in the user interface (UI).
-        /// </summary>
-        /// <returns>
-        ///     visibility definition of the selectable date-time-parts. The default is <see cref="TimePartVisibility.All" />.
-        /// </returns>
-        [Category("Appearance")]
-        [DefaultValue(TimePartVisibility.All)]
-        public TimePartVisibility PickerVisibility
-        {
-            get { return (TimePartVisibility)GetValue(PickerVisibilityProperty); }
-            set { SetValue(PickerVisibilityProperty, value); }
-        }
-
-        /// <summary>
         ///     Gets or sets the currently selected date.
         /// </summary>
         /// <returns>
@@ -301,97 +154,16 @@
             set { SetValue(SelectedDateProperty, value); }
         }
 
-        /// <summary>
-        ///     Gets or sets a collection used to generate the content for selecting the hours.
-        /// </summary>
-        /// <returns>
-        ///     A collection that is used to generate the content for selecting the hours. The default is a list of interger from 0
-        ///     to 23 if <see cref="IsMilitaryTime" /> is false or a list of interger from
-        ///     1 to 12 otherwise..
-        /// </returns>
-        [Category("Common")]
-        public ICollection<int> SourceHours
-        {
-            get { return (ICollection<int>)GetValue(SourceHoursProperty); }
-            set { SetValue(SourceHoursProperty, value); }
-        }
-
-        /// <summary>
-        ///     Gets or sets a collection used to generate the content for selecting the minutes.
-        /// </summary>
-        /// <returns>
-        ///     A collection that is used to generate the content for selecting the minutes. The default is a list of int from
-        ///     0 to 59.
-        /// </returns>
-        [Category("Common")]
-        public ICollection<int> SourceMinutes
-        {
-            get { return (ICollection<int>)GetValue(SourceMinutesProperty); }
-            set { SetValue(SourceMinutesProperty, value); }
-        }
-
-        /// <summary>
-        ///     Gets or sets a collection used to generate the content for selecting the seconds.
-        /// </summary>
-        /// <returns>
-        ///     A collection that is used to generate the content for selecting the minutes. The default is a list of int from
-        ///     0 to 59.
-        /// </returns>
-        [Category("Common")]
-        public ICollection<int> SourceSeconds
-        {
-            get { return (ICollection<int>)GetValue(SourceSecondsProperty); }
-            set { SetValue(SourceSecondsProperty, value); }
-        }
-
-        private CultureInfo SpecificCultureInfo
-        {
-            get { return Culture ?? Language.GetSpecificCulture(); }
-        }
+      
 
         public override void OnApplyTemplate()
         {
+            _calendar = GetTemplateChild(ElementCalendar) as System.Windows.Controls.Calendar;
             base.OnApplyTemplate();
-
-            UnSubscribeEvents();
-
-            _popup = GetTemplateChild(ElementPopup) as Popup;
-            _button = GetTemplateChild(ElementButton) as Button;
-            _hourInput = GetTemplateChild(ElementHourPicker) as Selector;
-            _minuteInput = GetTemplateChild(ElementMinutePicker) as Selector;
-            _secondInput = GetTemplateChild(ElementSecondPicker) as Selector;
-            _hourHand = GetTemplateChild(ElementHourHand) as FrameworkElement;
-            _ampmSwitcher = GetTemplateChild(ElementAmPmSwitcher) as Selector;
-            _minuteHand = GetTemplateChild(ElementMinuteHand) as FrameworkElement;
-            _secondHand = GetTemplateChild(ElementSecondHand) as FrameworkElement;
             _calendar = GetTemplateChild(ElementCalendar) as System.Windows.Controls.Calendar;
 
-            SetHandVisibility(HandVisibility);
-            SetPickerVisibility(PickerVisibility);
-
-            SetDefaultTimeOfDayValues();
-            SubscribeEvents();
-            ApplyCulture();
-            ApplyBindings();
         }
 
-        protected virtual void OnClosed()
-        {
-            RoutedEventHandler handler = Closed;
-            if (null != handler)
-            {
-                handler(this, new RoutedEventArgs());
-            }
-        }
-
-        protected virtual void OnOpened()
-        {
-            RoutedEventHandler handler = Opened;
-            if (null != handler)
-            {
-                handler(this, new RoutedEventArgs());
-            }
-        }
 
         protected virtual void OnSelectedDateChanged()
         {
@@ -407,63 +179,14 @@
             return Orientation.Vertical;
         }
 
-        private static object CoerceSource60(DependencyObject d, object basevalue)
-        {
-            var list = basevalue as ICollection<int>;
-            if (list != null)
-            {
-                return list.Where(i => i >= 0 && i < 60);
-            }
-            return null;
-        }
+       
 
-        private static object CoerceSourceHours(DependencyObject d, object basevalue)
-        {
-            var dt = d as DateTimePicker;
-            var hourList = basevalue as ICollection<int>;
-            if (dt != null && !dt.IsMilitaryTime && hourList != null)
-            {
-                return new Collection<int>(hourList.Where(i => i > 0 && i <= 12).OrderBy(i => i, new AmPmComparer()).ToList());
-            }
-            return basevalue;
-        }
-
-        private static bool IsValueSelected(Selector selector)
-        {
-            return selector != null && selector.SelectedItem != null;
-        }
 
         private static void OnClockVisibilityChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
             d.CoerceValue(OrientationProperty);
         }
 
-        private static void OnCultureChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            ((DateTimePicker)d).ApplyCulture();
-        }
-
-        private static void OnHandVisibilityChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            ((DateTimePicker)d).SetHandVisibility((TimePartVisibility)e.NewValue);
-        }
-
-        private static void OnIsDropDownOpenChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            if ((bool)e.NewValue)
-            {
-                ((DateTimePicker)d).OnOpened();
-            }
-            else
-            {
-                ((DateTimePicker)d).OnClosed();
-            }
-        }
-
-        private static void OnPickerVisibilityChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            ((DateTimePicker)d).SetPickerVisibility((TimePartVisibility)e.NewValue);
-        }
 
         private static void OnSelectedDateChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
@@ -472,35 +195,10 @@
             dateTimePicker.OnSelectedDateChanged(dt);
         }
 
-        private static void SetDefaultTimeOfDayValue(Selector selector)
+        protected override void ApplyBindings()
         {
-            if (selector != null)
-            {
-                if (selector.SelectedValue == null)
-                {
-                    selector.SelectedIndex = 0;
-                }
-            }
-        }
+            base.ApplyBindings();
 
-        private static void SetVisibility(UIElement partHours, UIElement partMinutes, UIElement partSeconds, TimePartVisibility visibility)
-        {
-            if (partHours != null)
-            {
-                partHours.Visibility = visibility.HasFlag(TimePartVisibility.Hour) ? Visibility.Visible : Visibility.Collapsed;
-            }
-            if (partMinutes != null)
-            {
-                partMinutes.Visibility = visibility.HasFlag(TimePartVisibility.Minute) ? Visibility.Visible : Visibility.Collapsed;
-            }
-            if (partSeconds != null)
-            {
-                partSeconds.Visibility = visibility.HasFlag(TimePartVisibility.Second) ? Visibility.Visible : Visibility.Collapsed;
-            }
-        }
-
-        private void ApplyBindings()
-        {
             if (_calendar != null)
             {
                 _calendar.SetBinding(System.Windows.Controls.Calendar.DisplayDateProperty, GetBinding(DisplayDateProperty));
@@ -510,148 +208,24 @@
                 _calendar.SetBinding(System.Windows.Controls.Calendar.IsTodayHighlightedProperty, GetBinding(IsTodayHighlightedProperty));
                 _calendar.SetBinding(FlowDirectionProperty, GetBinding(FlowDirectionProperty));
             }
-            if (_popup != null)
-            {
-                _popup.SetBinding(Popup.IsOpenProperty, GetBinding(IsDropDownOpenProperty));
-            }
         }
 
-        private void ApplyCulture()
+        protected sealed override void ApplyCulture()
         {
-            _deactivateRangeBaseEvent = true;
-            if (_ampmSwitcher != null)
-            {
-                _ampmSwitcher.Items.Clear();
-                if (!string.IsNullOrEmpty(SpecificCultureInfo.DateTimeFormat.AMDesignator))
-                {
-                    _ampmSwitcher.Items.Add(SpecificCultureInfo.DateTimeFormat.AMDesignator);
-                }
-                if (!string.IsNullOrEmpty(SpecificCultureInfo.DateTimeFormat.PMDesignator))
-                {
-                    _ampmSwitcher.Items.Add(SpecificCultureInfo.DateTimeFormat.PMDesignator);
-                }
-            }
+            base.ApplyCulture();
+
             if (_calendar != null)
             {
                 _calendar.Language = XmlLanguage.GetLanguage(SpecificCultureInfo.IetfLanguageTag);
             }
-            SetAmPmVisibility();
 
-            CoerceValue(SourceHoursProperty);
-            FirstDayOfWeek = SpecificCultureInfo.DateTimeFormat.FirstDayOfWeek;
-            if (SelectedDate.HasValue)
-            {
-                SetHourPartValues(SelectedDate.Value.TimeOfDay);
-            }
-            SetDefaultTimeOfDayValues();
-            _deactivateRangeBaseEvent = false;
+            FirstDayOfWeek = this.SpecificCultureInfo.DateTimeFormat.FirstDayOfWeek;
         }
 
-        /// <summary>
-        ///     Gets the offset from the selected <paramref name="currentHour" /> to use it in <see cref="TimeSpan" /> as hour
-        ///     parameter.
-        /// </summary>
-        /// <param name="currentHour"></param>
-        /// <returns>
-        ///     An integer representing the offset to add to the hour that is selected in the hour-picker for setting the correct
-        ///     <see cref="DateTime.TimeOfDay" />. The offset is determined as follows:
-        ///     <list type="table">
-        ///         <listheader>
-        ///             <term>Condition</term><description>Offset</description>
-        ///         </listheader>
-        ///         <item>
-        ///             <term><see cref="IsMilitaryTime" /> is false</term><description>0</description>
-        ///         </item>
-        ///         <item>
-        ///             <term>Selected hour is between 1 AM and 11 AM</term><description>0</description>
-        ///         </item>
-        ///         <item>
-        ///             <term>Selected hour is 12 AM</term><description>-12h</description>
-        ///         </item>
-        ///         <item>
-        ///             <term>Selected hour is between 12 PM and 11 PM</term><description>+12h</description>
-        ///         </item>
-        ///     </list>
-        /// </returns>
-        private int GetAmPmOffset(int currentHour)
+        protected override void OnRangeBaseValueChanged(object sender, SelectionChangedEventArgs e)
         {
-            if (!IsMilitaryTime)
-            {
-                if (currentHour == 12)
-                {
-                    if (GetSelectedHourDesignator() == SpecificCultureInfo.DateTimeFormat.AMDesignator)
-                    {
-                        return -12;
-                    }
-                }
-                else if (GetSelectedHourDesignator() == SpecificCultureInfo.DateTimeFormat.PMDesignator)
-                {
-                    return 12;
-                }
-            }
-            return 0;
-        }
-
-        private Binding GetBinding(DependencyProperty property)
-        {
-            return new Binding(property.Name) { Source = this };
-        }
-
-        private DateTime GetCorrectDateTime()
-        {
-            return SelectedDate.GetValueOrDefault(DateTime.Today).Date + GetTimeOfDay();
-        }
-
-        private string GetSelectedHourDesignator()
-        {
-            if (_ampmSwitcher.SelectedItem == null)
-            {
-                return null;
-            }
-            return (string)_ampmSwitcher.SelectedItem;
-        }
-
-        private TimeSpan GetTimeOfDay()
-        {
-            if (_timeOfDayChanged || !_timeOfDay.HasValue)
-            {
-                if (IsValueSelected(_hourInput) &&
-                    IsValueSelected(_minuteInput) &&
-                    IsValueSelected(_secondInput))
-                {
-                    var hours = (int)_hourInput.SelectedItem;
-                    var minutes = (int)_minuteInput.SelectedItem;
-                    var seconds = (int)_secondInput.SelectedItem;
-
-                    hours += GetAmPmOffset(hours);
-
-                    _timeOfDay = new TimeSpan(hours, minutes, seconds);
-                }
-                else
-                {
-                    _timeOfDay = SelectedDate.GetValueOrDefault(DateTime.Today).TimeOfDay;
-                }
-                _timeOfDayChanged = false;
-            }
-            return _timeOfDay.Value;
-        }
-
-        private void OnButtonClicked(object sender, RoutedEventArgs e)
-        {
-            IsDropDownOpen = !IsDropDownOpen;
-            if (_popup != null)
-            {
-                _popup.IsOpen = IsDropDownOpen;
-            }
-        }
-
-        private void OnRangeBaseValueChanged(object sender, SelectionChangedEventArgs e)
-        {
-            if (_deactivateRangeBaseEvent)
-            {
-                return;
-            }
-            _timeOfDayChanged = true;
+            base.OnRangeBaseValueChanged(sender, e);
+            
             SetDatePartValues();
         }
 
@@ -678,20 +252,12 @@
             }
         }
 
-        private void SetAmPmVisibility()
+
+        private DateTime GetCorrectDateTime()
         {
-            if (_ampmSwitcher != null)
-            {
-                if (!PickerVisibility.HasFlag(TimePartVisibility.Hour))
-                {
-                    _ampmSwitcher.Visibility = Visibility.Collapsed;
-                }
-                else
-                {
-                    _ampmSwitcher.Visibility = IsMilitaryTime ? Visibility.Collapsed : Visibility.Visible;
-                }
-            }
+            return SelectedDate.GetValueOrDefault(DateTime.Today).Date + GetTimeOfDay();
         }
+
 
         private void SetDatePartValues()
         {
@@ -703,100 +269,26 @@
             }
         }
 
-        private void SetDefaultTimeOfDayValues()
+        protected override void SubscribeEvents()
         {
-            SetDefaultTimeOfDayValue(_hourInput);
-            SetDefaultTimeOfDayValue(_minuteInput);
-            SetDefaultTimeOfDayValue(_secondInput);
-            SetDefaultTimeOfDayValue(_ampmSwitcher);
-        }
+            base.SubscribeEvents();
 
-        private void SetHandVisibility(TimePartVisibility visibility)
-        {
-            SetVisibility(_hourHand, _minuteHand, _secondHand, visibility);
-        }
-
-        private void SetHourPartValues(TimeSpan timeOfDay)
-        {
-            _deactivateRangeBaseEvent = true;
-            if (_hourInput != null)
-            {
-                if (!IsMilitaryTime)
-                {
-                    _ampmSwitcher.SelectedValue = timeOfDay.Hours < 12 ? SpecificCultureInfo.DateTimeFormat.AMDesignator : SpecificCultureInfo.DateTimeFormat.PMDesignator;
-                    if (timeOfDay.Hours == 0 || timeOfDay.Hours == 12)
-                    {
-                        _hourInput.SelectedValue = 12;
-                    }
-                    else
-                    {
-                        _hourInput.SelectedValue = timeOfDay.Hours % 12;
-                    }
-                }
-                else
-                {
-                    _hourInput.SelectedValue = timeOfDay.Hours;
-                }
-            }
-            if (_minuteInput != null)
-            {
-                _minuteInput.SelectedValue = timeOfDay.Minutes;
-            }
-            if (_secondInput != null)
-            {
-                _secondInput.SelectedValue = timeOfDay.Seconds;
-            }
-            _deactivateRangeBaseEvent = false;
-        }
-
-        private void SetPickerVisibility(TimePartVisibility visibility)
-        {
-            SetVisibility(_hourInput, _minuteInput, _secondInput, visibility);
-            SetAmPmVisibility();
-        }
-
-        private void SubscribeEvents()
-        {
-            SubscribeRangeBaseValueChanged(_hourInput, _minuteInput, _secondInput, _ampmSwitcher);
-
-            if (_button != null)
-            {
-                _button.Click += OnButtonClicked;
-            }
             if (_calendar != null)
             {
                 _calendar.SelectedDatesChanged += OnSelectedDatesChanged;
             }
         }
 
-        private void SubscribeRangeBaseValueChanged(params Selector[] selectors)
-        {
-            foreach (var selector in selectors.Where(i => i != null))
-            {
-                selector.SelectionChanged += OnRangeBaseValueChanged;
-            }
-        }
 
-        private void UnSubscribeEvents()
+        protected override void UnSubscribeEvents()
         {
-            UnsubscribeRangeBaseValueChanged(_hourInput, _minuteInput, _secondInput, _ampmSwitcher);
+            base.UnSubscribeEvents();
 
-            if (_button != null)
-            {
-                _button.Click -= OnButtonClicked;
-            }
             if (_calendar != null)
             {
                 _calendar.SelectedDatesChanged -= OnSelectedDatesChanged;
             }
         }
 
-        private void UnsubscribeRangeBaseValueChanged(params Selector[] selectors)
-        {
-            foreach (var selector in selectors.Where(i => i != null))
-            {
-                selector.SelectionChanged -= OnRangeBaseValueChanged;
-            }
-        }
     }
 }
