+++ MahApps.Metro/MahApps.Metro.csproj
@@ -171,6 +171,7 @@
     <Compile Include="Controls\SplitButton.cs" />
     <Compile Include="Controls\Theme.cs" />
     <Compile Include="Controls\TimePartVisibility.cs" />
+    <Compile Include="Controls\TimePicker.cs" />
     <Compile Include="Controls\WindowButtonCommands.cs" />
     <Compile Include="Controls\WindowCommands.cs">
       <SubType>Code</SubType>
